refactor: move public chrome from root to _app layout
