+++ app/root.tsx
@@ -6,25 +6,15 @@ import './styles/app.css';
 import faviconIcoUrl from './assets/images/favicon.ico';
 import icon32Url from './assets/images/icon-32x32.png';
 import appleTouchIconUrl from './assets/images/apple-touch-icon.png';
-import { backgroundImageStyles } from './styles/background-image';
-import { Breadcrumbs } from './navigation/breadcrumbs';
 import { appName, appThemeColor } from './core/config';
-import { SkipNavLink } from './core/skip-nav-link';
-import { Header } from './core/header';
-import { Footer } from './core/footer';
 import { securityHeaders } from './middleware/security-headers';
 import { loggingMiddleware } from './middleware/logging';
 import { getEnvVars } from './utils/env.server';
-import type { AppRouteHandle } from './types';
 import type { Route } from './+types/root';
 
 const { COMMIT_SHA } = getEnvVars();
 
 export const middleware: Route.MiddlewareFunction[] = [loggingMiddleware, securityHeaders];
-
-export const handle: AppRouteHandle = {
-  breadcrumb: () => ({ title: 'All Drinks' }),
-};
 
 export const links: LinksFunction = () => [
   {
@@ -55,11 +45,7 @@ export default function App({ loaderData }: Route.ComponentProps) {
   }, []);
 
   return (
-    <html
-      lang="en"
-      className="m-0 min-h-screen bg-neutral-800 bg-cover bg-fixed bg-center bg-no-repeat p-0 leading-tight"
-      data-commit={loaderData.commit}
-    >
+    <html lang="en" className="m-0 min-h-screen p-0 leading-tight" data-commit={loaderData.commit}>
       <head>
         <meta charSet="utf-8" />
         <meta name="viewport" content="width=device-width,initial-scale=1" />
@@ -71,18 +57,9 @@ export default function App({ loaderData }: Route.ComponentProps) {
         <meta name="twitter:card" content="summary_large_image" />
         <Meta />
         <Links />
-        <style dangerouslySetInnerHTML={{ __html: backgroundImageStyles }} />
       </head>
       <body className="relative flex min-h-screen flex-col font-sans font-light">
-        <SkipNavLink contentId="main" />
-        <Header />
-        <div className="flex flex-1 flex-col gap-6 py-4 sm:w-104 sm:gap-8 sm:self-center sm:py-8 lg:w-full lg:max-w-240 xl:max-w-7xl">
-          <Breadcrumbs />
-          <main id="main">
-            <Outlet />
-          </main>
-        </div>
-        <Footer />
+        <Outlet />
         <ScrollRestoration />
         <Scripts />
       </body>
