+++ app/styles/background-image.ts
@@ -2,12 +2,12 @@ import bgImg from '#/app/assets/images/background-768.jpg';
 import bgImgLg from '#/app/assets/images/background-2078.jpg';
 
 export const backgroundImageStyles = `
-  html {
+  .bg-app-image {
     background-image: url('${bgImg}');
   }
 
   @media (min-width: 1024px) {
-    html {
+    .bg-app-image {
       background-image: url('${bgImgLg}');
     }
   }
